+++ addons/ueipab_attendance_report/wizard/__init__.py
@@ -1,2 +1,3 @@
 from . import hr_attendance_report_wizard
 from . import vote_assist_wizard
+from . import hr_attendance_revision_wizard